+++ mantis-common/src/test/java/io/mantisrx/common/AckTest.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2022 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.mantisrx.common;
+
+import static org.junit.Assert.assertEquals;
+
+import org.junit.Test;
+
+public class AckTest {
+    private final JsonSerializer serializer = new JsonSerializer();
+
+    @Test
+    public void testAckInstance() throws Exception {
+        Ack ack = Ack.getInstance();
+        String serialized = serializer.toJson(ack);
+        Ack actual = serializer.fromJSON(serialized, Ack.class);
+        assertEquals(ack, actual);
+    }
+}
